JVM: move loadCompiledInlineFunction out of InlineCodegen
